# LibJS: Performance improvements around the `CreatePerIterationEnvironment` AO