@trflynn89 2022-03-08
> Typo in multiple places: DeclaritiveEnvironment -> DeclarativeEnvironment

I think I've been spelling this wrong my whole life :grimacing: fixed!